+++ packages/auth/__tests__/auth-creds-test.ts
@@ -3,11 +3,15 @@ import { Credentials } from '@aws-amplify/core';
 import { AuthOptions } from '../src/types';
 import {
 	CognitoUser,
-	CognitoUserPool,
 	CognitoUserSession,
 	CognitoAccessToken,
 	CognitoIdToken,
+	CognitoUserPool,
 } from 'amazon-cognito-identity-js';
+import {
+	InternalCognitoUser,
+	InternalCognitoUserPool,
+} from 'amazon-cognito-identity-js/internals';
 const authOptions: AuthOptions = {
 	userPoolId: 'us-west-2_0xxxxxxxx',
 	userPoolWebClientId: 'awsUserPoolsWebClientId',
@@ -21,7 +25,7 @@ describe('credentials syncing tests', () => {
 		const auth = new Auth(authOptions);
 
 		jest
-			.spyOn(CognitoUser.prototype, 'authenticateUser')
+			.spyOn(InternalCognitoUser.prototype, 'authenticateUser')
 			.mockImplementation((authenticationDetails, callback) => {
 				const session = new CognitoUserSession({
 					AccessToken: new CognitoAccessToken({ AccessToken: 'accesstoken' }),
@@ -32,7 +36,7 @@ describe('credentials syncing tests', () => {
 			});
 
 		jest
-			.spyOn(CognitoUserPool.prototype, 'getCurrentUser')
+			.spyOn(InternalCognitoUserPool.prototype, 'getCurrentUser')
 			.mockImplementation(() => {
 				return new CognitoUser({
 					Pool: new CognitoUserPool({
@@ -49,7 +53,7 @@ describe('credentials syncing tests', () => {
 		});
 
 		jest
-			.spyOn(CognitoUser.prototype, 'getSession')
+			.spyOn(InternalCognitoUser.prototype, 'getSession')
 			.mockImplementation((callback: any) => {
 				callback(null, session);
 			});