+++ packages/auth/src/providers/cognito/tokenProvider/cacheTokens.ts
@@ -1,3 +1,5 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
 import { AmplifyError, decodeJWT } from '@aws-amplify/core';
 import { AuthenticationResultType } from '@aws-sdk/client-cognito-identity-provider';
 import { tokenOrchestrator } from '.';