+++ packages/storage/src/common/S3ClientUtils.ts
@@ -74,14 +74,13 @@ interface S3InputConfig {
 	useAccelerateEndpoint?: boolean;
 	abortSignal?: AbortSignal;
 	emitter?: EventEmitter;
-	storageAction: StorageAction;
+	userAgentValue?: string;
 	dangerouslyConnectToHttpEndpointForTesting?: boolean;
 }
 
 export interface S3ResolvedConfig
 	extends Omit<S3InputConfig, 'region' | 'credentials'> {
 	region: string;
-	userAgentValue?: string;
 	credentials: () => Promise<AwsCredentials>;
 	customEndpoint?: string;
 	forcePathStyle?: boolean;
@@ -91,7 +90,7 @@ export interface S3ResolvedConfig
  * A function that persists the s3 configs, so we don't need to
  * assign each config parameter for every s3 API call.
  *
- * @inernal
+ * @internal
  */
 export const loadS3Config = (config: S3InputConfig): S3ResolvedConfig => {
 	if (!config.region) {
@@ -104,10 +103,6 @@ export const loadS3Config = (config: S3InputConfig): S3ResolvedConfig => {
 		credentials: config.credentials
 			? () => Promise.resolve(config.credentials!)
 			: credentialsProvider,
-		userAgentValue: getAmplifyUserAgent({
-			category: Category.Storage,
-			action: config.storageAction,
-		}),
 		...(config.dangerouslyConnectToHttpEndpointForTesting
 			? {
 					customEndpoint: localTestingStorageEndpoint,