+++ packages/geo/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.3.1 (2023-08-10)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+# [2.3.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@2.2.0...@aws-amplify/geo@2.3.0) (2023-07-31)
+
+### Features
+
+- custom user agent Geo changes for UI handoff ([#11632](https://github.com/aws-amplify/amplify-js/issues/11632)) ([01bfa8f](https://github.com/aws-amplify/amplify-js/commit/01bfa8f692737bd14422f7dc2eae11ed00c19048))
+
+# [2.2.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@2.1.4...@aws-amplify/geo@2.2.0) (2023-07-20)
+
+### Features
+
+- custom user agent core changes for UI handoff ([#11602](https://github.com/aws-amplify/amplify-js/issues/11602)) ([7365c34](https://github.com/aws-amplify/amplify-js/commit/7365c34b28015af199dbfdb3713cc26e096d1213))
+
+## [2.1.4](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@2.1.3...@aws-amplify/geo@2.1.4) (2023-07-13)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
 ## [2.1.3](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@2.1.2...@aws-amplify/geo@2.1.3) (2023-06-28)
 
 **Note:** Version bump only for package @aws-amplify/geo